requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
gspread>=6.1.0
google-auth>=2.29.0
playwright>=1.44.0
pypdf>=4.2.0
python-dotenv>=1.0.1
